+++ MyEssentialFeed/MyEssentialFeedTests/Feed API/URLSesstionHTTPClientTests.swift
@@ -124,16 +124,8 @@ final class URLSesstionHTTPClientTests: XCTestCase {
         return receivedResult!
     }
     
-    private func anyURL() -> URL {
-        return URL(string: "http://any-url.com")!
-    }
-    
     private func anyData() -> Data {
         return Data("any".utf8)
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any error", code: 0)
     }
     
     private func nonHTTPURLResponse() -> URLResponse {
